requirements.txt
pika == 1.3.1
requests == 2.28.2
bs4 == 0.0.1
python-dotenv == 0.21.1
pyyaml == 6.0

beautifulsoup4~=4.11.1
urllib3~=1.26.5
cloudscraper~=1.2.66